Refactor container_test.ts out from topology-test.ts
